move remaining print funcs out
