+++ src/pages/List/CardSelector/index.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useReducer } from 'react';
 import Fade from '@material-ui/core/Fade';
 import Typography from '@material-ui/core/Typography';
 import ListContext from 'context/ListContext';
